Only show export button when messages are visible
